requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
sqlparse==0.5.3
django-cors-headers==4.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.2.0
django-filter==24.3
drf-yasg==1.21.10